requirements.txt
Django>=1.8
django-ace-overlay==0.5
html2text==2016.5.29